requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.0
gunicorn==20.1.0
numpy==1.21.4
pandas==1.3.5
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
unicorn==1.0.3
whitenoise==5.3.0
